Fix Rakefile to not modify ENV
Instead of modifying ENV and running code inside the rake process,
run a ruby process with the environment set correctly.  This fixes:

  rake test_up dev_up

which previously would migrate the test database up and not the
development database up. Have the specs test that this and similar
task combinations work.

Since I had to modify it anyway, make the annotate task use the
test environment and not the development environment.  It's generally
problem-free to rebuild a test environment by migrating down and up,
while there are cases where that isn't great in a development
environment.

Update the stack specs to use separate databases for the development,
test, and production environments, and test that each of them work
as expected. Also test that the annotation is actually applied to the
model file.